requirements/base.txt
asgiref==3.6.0
Django==4.2.1
django-browser-reload==1.8.0
django-cors-headers==3.14.0
django-restframework==0.0.1
django-simple-history==3.3.0
django-tailwind==3.5.0
django-widget-tweaks==1.4.12
django-environ==0.10.0
djangorestframework==3.14.0
django-compressor==4.3.1
django-simple-captcha==0.5.17
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
watchdog==3.0.0
whitenoise==6.4.0
pymemcache==4.0.0
redis==4.5.5